[#188] refactor: presentation layer 폴더 정리
