+++ lib/common/routers/test_page.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:wehavit/common/common.dart';
-import 'package:wehavit/presentation/auth/presentation/providers/auth_provider.dart';
+
+import '../../presentation/auth/presentation.dart';
 
 class TestPage extends ConsumerWidget {
   const TestPage({super.key});
