--- lib/presentation/auth/data/entities/auth_result.dart
@@ -1,1 +1,0 @@
-// TODO Implement this library.
