+++ lib/presentation/auth/providers/auth_notifier.dart
@@ -5,7 +5,7 @@ import 'package:wehavit/domain/usecases/login/email_and_password_login_usecase.d
 import 'package:wehavit/domain/usecases/login/email_and_password_register_usecase.dart';
 import 'package:wehavit/domain/usecases/login/google_login_usecase.dart';
 import 'package:wehavit/domain/usecases/logout/logout_usecase.dart';
-import 'package:wehavit/presentation/auth/auth.dart';
+import 'package:wehavit/presentation/auth/presentation.dart';
 
 class AuthNotifier extends StateNotifier<AuthState> {
   AuthNotifier(
