+++ lib/presentation/auth/providers/auth_provider.dart
@@ -3,7 +3,7 @@ import 'package:wehavit/domain/usecases/login/email_and_password_login_usecase_p
 import 'package:wehavit/domain/usecases/login/email_and_password_register_usecase_provider.dart';
 import 'package:wehavit/domain/usecases/login/google_login_usecase_provider.dart';
 import 'package:wehavit/domain/usecases/logout/logout_usecase_provider.dart';
-import 'package:wehavit/presentation/auth/auth.dart';
+import 'package:wehavit/presentation/auth/presentation.dart';
 
 final authProvider = StateNotifierProvider<AuthNotifier, AuthState>((ref) {
   final usecaseGoogleLogIn = ref.watch(googleLogInUseCaseProvider);
