+++ lib/presentation/auth/providers/auth_state.dart
