+++ lib/presentation/auth/screens/auth_screen.dart
@@ -5,7 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:wehavit/common/common.dart';
 import 'package:wehavit/data/models/auth_result_model.dart';
-import 'package:wehavit/presentation/features.dart';
+import 'package:wehavit/presentation/auth/presentation.dart';
 
 class AuthScreen extends HookConsumerWidget {
   const AuthScreen({super.key});
