--- lib/presentation/features.dart
@@ -1,3 +1,0 @@
-export 'auth/auth.dart';
-export 'friend_list/friend_list.dart';
-export 'splash/splash.dart';
