--- lib/presentation/friend_list/friend_list.dart
@@ -1,1 +1,0 @@
-export 'presentation/presentation.dart';
