+++ lib/presentation/friend_list/screens/friend_list_screen.dart
@@ -1,11 +1,10 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:wehavit/common/common.dart';
-import 'package:wehavit/presentation/friend_list/presentation/providers/friend_list_provider.dart';
-import 'package:wehavit/presentation/friend_list/presentation/widgets/add_friend_textfield_widget.dart';
-import 'package:wehavit/presentation/friend_list/presentation/widgets/friend_element_widget.dart';
+import 'package:wehavit/presentation/friend_list/providers/friend_list_provider.dart';
+import 'package:wehavit/presentation/friend_list/widgets/add_friend_textfield_widget.dart';
+import 'package:wehavit/presentation/friend_list/widgets/friend_element_widget.dart';
 
 class FriendListScreen extends ConsumerStatefulWidget {
   const FriendListScreen({super.key});
