+++ lib/presentation/friend_list/widgets/add_friend_textfield_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wehavit/common/common.dart';
-import 'package:wehavit/presentation/friend_list/presentation/providers/add_friend_provider.dart';
-import 'package:wehavit/presentation/friend_list/presentation/providers/friend_list_provider.dart';
+import 'package:wehavit/presentation/friend_list/providers/add_friend_provider.dart';
+import 'package:wehavit/presentation/friend_list/providers/friend_list_provider.dart';
 
 class AddFriendTextFieldWidget extends ConsumerWidget {
   const AddFriendTextFieldWidget({super.key});
